authorization done

+++ Backend/routes/posts.js
@@ -35,7 +35,8 @@ router.post('',
         const post = new Post({
             title: req.body.title,
             content: req.body.content,
-            imagePath: url + '/images/' + req.file.filename
+            imagePath: url + '/images/' + req.file.filename,
+            creator: req.userData.userId
         });
         post.save().then(createdPost => {
             res.status(201).json({
@@ -60,11 +61,16 @@ router.put('/:id',
             _id: req.body.id,
             title: req.body.title,
             content: req.body.content,
-            imagePath: imagePath
+            imagePath: imagePath,
+            creator: req.userData.userId
         });
-        Post.updateOne({ _id: req.params.id }, post)
+        Post.updateOne({ _id: req.params.id, creator: req.userData.userId /* only the creator update his post */ }, post)
             .then(result => {
-                res.status(200).json({ message: 'Updated successeful!' });
+                if (result.nModified > 0) {
+                    res.status(200).json({ message: 'Updated successeful!' });
+                } else {
+                    res.status(401).json({ message: 'Not authorized!' });
+                }
             })
     });
 
@@ -107,10 +113,13 @@ router.get('/:id', (req, res, next) => {
 router.delete('/:id',
     checkAuth,
     (req, res, next) => {
-        Post.deleteOne({ _id: req.params.id })
+        Post.deleteOne({ _id: req.params.id, creator: req.userData.userId })
             .then(result => {
-                console.log(result);
-                res.status(200).json({ message: 'Post deleted!' });
+                if (result.n > 0) {
+                    res.status(200).json({ message: 'Post Deleted!' });
+                } else {
+                    res.status(401).json({ message: 'Not authorized!' });
+                }
             })
     });
